requirements.txt
python-telegram-bot[all]==20.6
pymongo
yt-dlp
requests